requirements copy.txt
# Backend dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10

# AI Integration
litellm==1.11.1
python-dotenv==1.0.0

# Frontend
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3

# Utilities
python-dateutil==2.8.2